Start row model init, move aggregation out of
helper, into util called delegate-aggregator;
