+++ app/templates/components/data-table-row-current.hbs
@@ -1,3 +1,5 @@
+{{log data.isSpecial}}
+
 {{#if rowconfig.divider}}
   <td>&nbsp;</td>
   <td class="cell-border-left">&nbsp;</td>
