+++ app/templates/components/data-table.hbs
@@ -25,7 +25,6 @@
     <table class="body-table">
       <tbody>
         {{#each config as |row|}}
-          {{log row.isSpecial}}
           {{#if hasBlock}}
             {{yield}}
           {{else}}
@@ -46,18 +45,8 @@
                 comparison=comparison
                 rowconfig=row
                 options=profile
-                data=
-                  (unless
-                    row.special
-                    (get this.t1 row.data)
-                    (aggregate-special-variable rowConfig=row data=t1)
-                  )
-                data2=
-                  (unless
-                    row.special
-                    (get this.t2 row.data)
-                    (aggregate-special-variable rowConfig=row data=t2)
-                  )
+                data=(get this.t1 row.data)
+                data2=(get this.t2 row.data)
               }}
             {{/if}}
           {{/if}}
