+++ tests/unit/utils/delegate-aggregator-test.js
@@ -1,0 +1,10 @@
+import delegateAggregator from 'labs-nyc-factfinder/utils/delegate-aggregator';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | delegate aggregator');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = delegateAggregator();
+  assert.ok(result);
+});
